Create coke.py
This program is like a cashier for a coke machine, it prompts user to insert coins, which is supposed to be of some particular denomination. The machine ignores every other coin denomination, except for it accepted denomination, and it also calculates the amount due, and change owed, once user inserts coin.