Create Felipe's Taqueria.py
This problem set tasks the program with prompting for, and taking user's inputs. In this case, an order or sequence of orders, off the menu of Felipe's Taqueria is given by the user as input, the total price is updated after each item is ordered, and the order is conclusive when control-d is inputted.