Create Adieu, adieu.py
The adieu problem set requires the program to continuously prompt user for input, until ctrl-d to signal end of input. The inputs which are expected to be names, are then greeted adieu with commas(,) and, and placed at the right places in between names, with the help of the inflect module.