Create Guess game.py
This problem set requires the program to prompt user for input of a positive integer which is stored as level, the program is expected to generate a random integer within the range of 1 and the integer stored as level, inclusive. The program then prompts user for a guess which  is compared with the random generated integer.