Create Little professor.py
This problem set requires the program to take user's input in the form of a level, the program is expected to only accept an input of either 1, 2, or 3, and is expected to re-prompt if any other number or non digit is inputted. Based on level input, a set of random mathematical equations are outputted, which user is expected to answer, now on every wrong answer, "EEE" is printed out, and same equation prompted until after wrong answer is given 3 times, prompting the program to produce the correct answer, then a new equation pops up. Now upon input of correct answer, a new equation is outputted till a total of 10 equations are attempted after which a total score is printed out.