Create Pizza.py
This problem set requires the program to only accept a single command-line argument, check if the command-line argument is an existing csv file, then the csv file is opened, read, and then tabulated in grid format. NOTE: The headers are read using next(), in order to achieve the double dotted grid effect that demarcates the headers from the items on the list.