Create Watch.py
The watch.py problem set requires the program to call the parse function on the user's input. The parse function searches an inputted HTML format, youtube URL, using re.search, and if matches are found, the youtube link equivalent is returned as output, whereas None is returned as output in the absence of a match.