Create seasons.py
The seasons.py problem set requires the program to take user's input of date of birth, in the format of YY-MM-DD, and would exit, or quit if any other format of D.O.B. is inputted. The program then calculates the user's age in minutes, via the D.O.B passed as input, and returns as output said age in words without any "and".